Add files via upload
When guessing a word, if there is 2 of the same letter in one guess, and one of the letters is green or orange, the corresponding key on the keyboard below will be green or orange and not gray. Also added average guesses on statistics, which tell you the average amount of guesses you take to either guess the word or run out after 6 guesses.