Add files via upload
when user types in a word that is not found in the system, the red message telling the user the word is not found will not appear if the message is already there. If it's not there, then it will appear to tell the user. The word continues shaking from side to side if you the word is not found.